requirements.txt
streamlit==1.34.1
ultralytics==8.2.2
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.9.0.80
Pillow==10.2.0
